fix(typography): apply font CSS variables to <html> via _document.tsx
After migrating from App Router to Pages Router, the font CSS variables
(--font-playfair-display, --font-lora) were only scoped to <main>, so
globals.css rules on body, Header, and Footer couldn't resolve them.

Moves font definitions to _document.tsx and applies the variable classes
to <Html>, making them available at the root level for all selectors.
